move test to unit test folder
